third_party/blink/web_tests/flag-specific/highdpi/fast/css/round-up-all-line-width-less-than-1-greater-than-0-expected.txt

This is a testharness.js-based test.
FAIL Test that chrome rounds up all borders/outlines when they are less than 1px but greater than 0px. assert_equals: expected "1px" but got "0.666667px"
Harness: the test ran to completion.
